Added a __name__ field
The __title__ field that a lot of Python packages have adopted seems
to be used for the name of the package, which in our case is "holdit".
I needed the real name ("Holdit!")  so that I could put it in titles.
I think that's what really should be the __title__ and __name__ should
have been used for the package name, but it would probably be
confusing to other people to go against convention.  So, I used
__name__ for "Holdit!".