Move LoginDialog code to control.py and use pubsub for communication
The previous implementation worked on macOS but hung on Windows.  I'm
still not entirely clear why, but it definitely had something to do
with putting the wait queue on the wrong object.  The new approach now
puts the login dialog code in with the main control gui, and uses
WxPython's publish-subscribe functionality to send messages to invoke
the login dialog.  This appears to work on Windows and no longer
causes everything to hang up.