New implementation of a credentials storage module
The original implementation of oauth2client.client.token_storage stores
the credentials from Google as the value of the password in a call to
keyring.set_password(...).  Unfortunately, this fails on Windows because of
size limitations on the length of passwords.  The error is obscure and
unobvious.

This implementation stores an encryption key in the password field instead,
and uses that encryption key to encrypt a file that stores the credentials
on disk.  This is not very secure, but for our in-house application for the
Caltech Library, it's probably good enough.